Use pre-defined unofficial title IDs for some homebrew
This is to keep a specific title ID throughout each release's bundled prebuilt NSPs.

It also helps the community sort of decide on a title ID when making and sharing updated NSP forwarders (e.g., to update the version listed under Title Information, or a new icon comes out).

With a static icon NSPs can be installed to overwrite and remove the old version which helps as well.