Don't require the ROM to exist
Unlike the NRO, we never need to read anything from it. The only downside with this change is it will not be able to check if the ROM path provided actually exists.

The benefit though is you can more easily batch and script ROM paths without needing the Switch's microSD card to be inserted to the PC.