Allow manually specifying the sdmc NRO path with --sdmc
This allows the user to not require the NRO path to be on the Switch's microSD card and can instead be anywhere on their PC or any other storage.

The --sdmc MUST be the full path to where the NRO file is on the Switch's microSD card, excluding any drive letter. You don't need to start it with `sdmc:/`.
This is effectively the path to the NRO file that the NSP will load.

The NRO path must still exist and be an NRO file. It is expected to be the same NRO file that the manual --sdmc specifies.