Add --no-verify to skip NRO verification as a last-resort
This should only be used if nstool has some software incompatibility issue with a specific NRO. You should take extra care that the NRO is not corrupt.

Note that if nstool failed to verify the NRO, it will likely fail for all other actions with the NRO like NACP and Icon extraction as well.