Require a-zA-Z0-9 at least once in the Title and Publisher
This is to stop building NSPs with pretty vague details like '-' as a publisher name for DBI as an example.
It's intentionally a somewhat loose check so that if someone really did have some special characters in funky ways, they should be fine as long as theres at least one 'a' or such for example.